requirements.txt
python-telegram-bot==22.7
mysql-connector-python==8.4.0
openpyxl==3.1.5
reportlab==4.2.2
